Add files via upload
The most widely used open-source container orchestration engine in the world is Kubernetes. It provides scalable scheduling and management of containers, whether they are running on Docker or not. Learn the fundamentals of Kubernetes in this course to prepare for your first trip. an overview of the cloud-native ecosystem and the reasons behind the use of Kubernetes by hundreds of additional businesses, like Spotify and Wikimedia. how to use security best practises, deploy a containerized application and manage it with Kubernetes, establish a Kubernetes cluster with Minikube, and more. You will engage in activities that provide you practical experience along the way.